+++ NEWS.md
@@ -11,7 +11,7 @@ Changes
 
 - Allow assigning query result to an Andromeda table where the query uses that Andromeda table. (e.g. `andromeda$cars <- andromeda$cars %>% filter(speed > 10)`)
 
-- Added `showProgressBar` arguments to `batchApply` and `groupApply`.
+- Added `progressBar` arguments to `batchApply` and `groupApply`.
 
 Bug fixes